Upgraded from an older, buggier version of Handlebars to the latest.
