+++ index.html
@@ -19,7 +19,7 @@
 	<link rel="stylesheet" href="css/style.css?v=2">
 	<link rel="stylesheet" media="handheld" href="css/handheld.css?v=2">
 	<script src="js/libs/modernizr-1.7.min.js"></script>
-	<script src="js/libs/handlebars-1.0.0.beta.2.js"></script>	
+	<script src="js/libs/handlebars-1.0.0.beta.6.js"></script>	
 		
 <script id ="getting-started-1" type="text/x-handlebars-template">
 <div class="entry">
